+++ lib/payrix/configuration.rb
@@ -3,18 +3,19 @@
 module Payrix
   # Use this class to configure API parameters such as API URL, API key, etc.
   class Configuration
-    attr_accessor :api_key, :session_key, :test_mode
+    attr_accessor :api_key, :session_key, :environment
 
     def initialize
       @api_key = ''
       @session_key = ''
-      @test_mode = false
+      @environment = Payrix::ENVIRONMENTS.fetch(:production)
     end
 
     def url
-      if @test_mode
+      case @environment
+      when Payrix::ENVIRONMENTS.fetch(:sandbox)
         'https://test-api.payrix.com'
-      else
+      when Payrix::ENVIRONMENTS.fetch(:production)
         'https://api.payrix.com'
       end
     end